Updates on couple of classes
I did some changes to Player and Subject to model more closely to
Observer pattern in the book. I have notifyObserver method that now
updates board base upon which notification method is called. If is a
move then it calls the moveUpdate method or if it is a place of a ship
it calls the shipPlacementUpdate method. In board I changed the
arrayList for ships_List and sunk_list to ship arrays because I felt
like we shouldn't have something that is suppose to grow in size. These
variables are suppose to be fixed.